requirements.txt
mkdocs-material~=8.5
pyyaml~=6.0
